Add files via upload
The "Market Sales Dataset" is a comprehensive collection of sales data that includes various features such as product categories, sales figures, timestamps, and more. This dataset is ideal for performing data analysis to uncover trends and insights in market sales.
- **Source**: Kaggle
- **File Format**: ZIP
- **Contents**: The ZIP file contains multiple CSV files, each representing different aspects of market sales data.

The dataset is stored in the `data` folder of this project. Upon extraction, the folder structure is as follows: